feat(react-property): consolidate injection logic to `index.js`
BREAKING CHANGE: remove exports `HTMLDOMPropertyConfig` and
`SVGDOMPropertyConfig` and consolidate the properties in
`properties`.

As a result of this change, `src/` directory is removed since all
the injection logic is handled in `index.js` and the npm script
`copy` is removed as well.
