+++ packages/react-property/package.json
@@ -4,11 +4,10 @@
   "description": "HTML and SVG DOM property configs used by React.",
   "main": "index.js",
   "scripts": {
-    "build": "npm run clean && npm run build:html && npm run build:injection && npm run build:svg && npm run copy",
+    "build": "npm run clean && npm run build:html && npm run build:injection && npm run build:svg",
     "build:html": "node scripts/build-html",
     "build:injection": "node scripts/build-injection",
     "build:svg": "node scripts/build-svg",
-    "copy": "cp -r src/* lib",
     "clean": "rm -rf lib",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
